Move SSL Wildcard Kluge to Flag
At one point activemq-cpp had a bug where it refused to validate wildcard
certificates.  The code I templated from had a line to disable SSL peer
validation to get around that bug - I'm not even sure if that bug is still
a thing.

In any case releasing something with that hard-coded isn't cool so I backed
it out to a flag.